backend/requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==8.4.0
importlib-resources==6.4.4
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.9.0
SQLAlchemy==2.0.32
typing-extensions==4.12.2
werkzeug==3.0.4
zipp==3.20.1
